Add files via upload
The best restaurants data sets are used for the people who are seeking for affordable and best restaurants during their vacation. It contains data sets of restaurants all over the world. And it has the solution set of those data in map view for clear understanding. And it has city wise and country wise restaurants solution set. And also we have found the nearby restaurants with lowest price.